# docs/07_api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zedly API v1",
    "version": "0.1.0",
    "description": "Executable API contract for current /api/v1 endpoints.",
    "contact": {
      "name": "Zedly API Team"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "authLogin",
        "tags": [
          "auth"
        ],
        "description": "Authenticate by login and password (OTP for first entry).",
        "security": [],
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeLoginResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/telegram": {
      "post": {
        "operationId": "authTelegram",
        "tags": [
          "auth"
        ],
        "description": "Authenticate via Telegram after account is linked in profile settings.",
        "security": [],
        "responses": {
          "200": {
            "description": "Telegram auth success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeTelegramAuth"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthTelegramRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "authRefresh",
        "tags": [
          "auth"
        ],
        "description": "Rotate refresh token and issue a new access token.",
        "security": [],
        "responses": {
          "200": {
            "description": "Refresh success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeTokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "tags": [
          "auth"
        ],
        "description": "Invalidate current refresh token.",
        "responses": {
          "200": {
            "description": "Logout success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeAuthStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLogoutRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "authLogoutAll",
        "tags": [
          "auth"
        ],
        "description": "Invalidate all active sessions for current user.",
        "responses": {
          "200": {
            "description": "Logout-all success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeAuthStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/change-first": {
      "post": {
        "operationId": "authChangeFirstPassword",
        "tags": [
          "auth"
        ],
        "description": "Complete mandatory OTP password change on first login.",
        "security": [],
        "responses": {
          "200": {
            "description": "Password changed and session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeFirstPasswordChange"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthFirstPasswordChangeRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "operationId": "authForgotPassword",
        "tags": [
          "auth"
        ],
        "description": "Request password reset token.",
        "security": [],
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeForgotPassword"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthForgotPasswordRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "example": {
          "code": "string",
          "message": "string"
        }
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "examples": [
          {
            "ok": false,
            "error": {
              "code": "resource_not_found",
              "message": "Entity not found",
              "details": {
                "resource": "report",
                "resource_id": "rpt_missing"
              }
            }
          }
        ]
      },
      "TokenPair": {
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in_seconds"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "bearer"
            ]
          },
          "expires_in_seconds": {
            "type": "integer",
            "minimum": 1
          }
        },
        "example": {
          "access_token": "string",
          "refresh_token": "string",
          "token_type": "bearer",
          "expires_in_seconds": 1
        }
      },
...
      "AuthStatusData": {
        "type": "object",
        "required": [
          "status",
          "sessions_terminated"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "sessions_terminated": {
            "type": "integer",
            "minimum": 0
          }
        },
        "example": {
          "status": "string",
          "sessions_terminated": 0
        }
      },
      "ForgotPasswordAcceptedData": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted"
            ]
          },
          "message": {
            "type": "string"
          },
          "debug_reset_token": {
            "type": "string"
          }
        },
        "example": {
          "status": "accepted",
          "message": "string"
        }
      },
...
      "PasswordChangeRequiredData": {
        "type": "object",
        "required": [
          "status",
          "challenge_token",
          "expires_in_seconds"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "password_change_required"
            ]
          },
          "challenge_token": {
            "type": "string"
          },
          "expires_in_seconds": {
            "type": "integer",
            "minimum": 1
          }
        },
        "example": {
          "status": "password_change_required",
          "challenge_token": "pwd_xxx",
          "expires_in_seconds": 900
        }
      },
      "LoginResultData": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/TokenPair"
          },
          {
            "$ref": "#/components/schemas/PasswordChangeRequiredData"
          }
        ]
      },
      "FirstPasswordChangeData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TokenPair"
          },
          {
            "type": "object",
            "required": [
              "status",
              "google_connect_url",
              "telegram_connect_url",
              "skip_label"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "login_methods_prompt"
                ]
              },
              "google_connect_url": {
                "type": "string"
              },
              "telegram_connect_url": {
                "type": "string"
              },
              "skip_label": {
                "type": "string"
              }
            }
          }
        ],
        "example": {
          "status": "login_methods_prompt",
          "google_connect_url": "/api/v1/users/me/login-methods/google/connect",
          "telegram_connect_url": "/api/v1/users/me/login-methods/telegram/connect",
          "skip_label": "Пропустить — сделаю позже",
          "access_token": "access.jwt.token",
          "refresh_token": "refresh.jwt.token",
          "token_type": "bearer",
          "expires_in_seconds": 900
        }
      },
...
      "TelegramNotConnectedData": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "telegram_not_connected"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": "telegram_not_connected",
          "message": "string"
        }
      },
      "TelegramAuthResultData": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/TokenPair"
          },
          {
            "$ref": "#/components/schemas/TelegramNotConnectedData"
          }
        ],
        "example": {
          "access_token": "string",
          "refresh_token": "string",
          "token_type": "bearer",
          "expires_in_seconds": 1
        }
      },
...
      "ApiSuccessEnvelopeTokenPair": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/TokenPair"
          }
        },
        "examples": [
          {
            "ok": true,
            "data": {
              "access_token": "access.jwt.token",
              "refresh_token": "refresh.jwt.token",
              "token_type": "bearer",
              "expires_in_seconds": 900
            }
          }
        ]
      },
      "ApiSuccessEnvelopeLoginResult": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/LoginResultData"
          }
        }
      },
      "ApiSuccessEnvelopeAuthStatus": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/AuthStatusData"
          }
        },
        "example": {
          "ok": true,
          "data": {
            "status": "string",
            "sessions_terminated": 0
          }
        }
      },
      "ApiSuccessEnvelopeForgotPassword": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/ForgotPasswordAcceptedData"
          }
        },
        "example": {
          "ok": true,
          "data": {
            "status": "accepted",
            "message": "string"
          }
        }
      },
      "ApiSuccessEnvelopeFirstPasswordChange": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/FirstPasswordChangeData"
          }
        }
      },
...
      "ApiSuccessEnvelopeTelegramAuth": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/TelegramAuthResultData"
          }
        },
        "example": {
          "ok": true,
          "data": {
            "access_token": "string",
            "refresh_token": "string",
            "token_type": "bearer",
            "expires_in_seconds": 1
          }
        }
      },
...
      "AuthLoginRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "login": "teacher.a.schoola.1",
          "password": "teacher-pass"
        }
      },
      "AuthTelegramRequest": {
        "type": "object",
        "required": [
          "auth_data"
        ],
        "properties": {
          "auth_data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "example": {
          "auth_data": {
            "id": "123456",
            "first_name": "Ali",
            "auth_date": "1700000000",
            "hash": "telegram_hash"
          }
        }
      },
      "AuthRefreshRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        },
        "example": {
          "refresh_token": "refresh.jwt.token"
        }
      },
      "AuthLogoutRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        },
        "example": {
          "refresh_token": "refresh.jwt.token"
        }
      },
      "AuthFirstPasswordChangeRequest": {
        "type": "object",
        "required": [
          "challenge_token",
          "new_password",
          "repeat_password"
        ],
        "properties": {
          "challenge_token": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          },
          "repeat_password": {
            "type": "string"
          }
        },
        "example": {
          "challenge_token": "pwd_abc123",
          "new_password": "student-pass-1",
          "repeat_password": "student-pass-1"
        }
      },
      "AuthForgotPasswordRequest": {
        "type": "object",
        "required": [
          "login"
        ],
        "properties": {
          "login": {
            "type": "string"
          }
        },
        "example": {
          "login": "teacher.a.schoola.1"
        }
      },
...
    }
  }
}